never override parameters provided explicitly by CAN-provided factors
The issue this fixes is in a case where the parameters are not stored
in the controller. In this case, some or all of the parameters are
passed explicitly with setMotorParameters

However, if for whatever reason some of the factor objects are read,
when received and processed by the controller, these factor objects
will reset the values set explicitly.

Change the behavior to always consider the explicit motor parameters
to be of a higher priority than the ones coming from the device.